requirements.txt
streamlit
numpy
mne
tensorflow
scikit-learn
matplotlib
pandas
pillow
